Exception handling - delete

+++ src/main/java/com/educandoweb/courseJavaNelio/services/UserService.java
@@ -4,10 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.educandoweb.courseJavaNelio.entities.User;
 import com.educandoweb.courseJavaNelio.repositories.UserRepository;
+import com.educandoweb.courseJavaNelio.services.exceptions.DataBaseException;
 import com.educandoweb.courseJavaNelio.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -29,7 +31,16 @@ public class UserService {
 	}
 	
 	public void delete (Long id) {
+		if(!repository.existsById(id)) {
+			throw new ResourceNotFoundException(id);
+		}
+		try {
 		repository.deleteById(id);
+		}
+		catch(DataIntegrityViolationException e) {
+			throw new DataBaseException(e.getMessage());
+		}
+		
 	}
 	
 	public User update(Long id, User obj) {
